backend/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.14
django-cors-headers>=4.3
mysqlclient>=2.2
pymysql>=1.1
transformers>=4.36
torch>=2.1
networkx>=3.2
numpy>=1.26
scikit-learn>=1.3
sentence-transformers>=2.3
requests>=2.31
python-dotenv>=1.0
cryptography>=42.0
